requirements.txt
certifi>=2025.4.26
charset-normalizer>=3.4.2
idna>=3.10
requests>=2.32.3
urllib3>=2.4.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
PyMuPDF>=1.25.0
python-docx>=1.1.0
rapidfuzz>=3.0.0
